// src/store/main/system/type.ts
/*
 * @Author: your name
 * @Date: 2021-08-25 20:59:55
 * @LastEditTime: 2021-08-31 20:19:35
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: /vue3-practice/src/store/main/system/type.ts
 */
export interface ISystemState {
  usersList: any[]
  usersCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  menuList: any[]
  menuCount: number
}